apps/backend/requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12,<4.0
psycopg2>=2.8,<3.0
gunicorn>=20.0,<21.0
python-dotenv>=0.15,<1.0
